explib/requirements.txt
torch==1.8.1
torchvision==0.9.1
wandb
pandas
numpy
scikit-learn>=0.24.2
tqdm
matplotlib
torchtext==0.9.1
Pillow>=0.0
statsmodels
transformers==4.9.1
accelerate
datasets >= 1.8.0
python-dotenv
